% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{adjust_time}
\alias{adjust_time}
\title{Adjust daily dive time.}
\usage{
adjust_time(
  dive_time,
  min_time = dive_pars$min_time,
  round_hr = dive_pars$round_hr
)
}
\arguments{
\item{dive_time}{Numeric. Dive time in minutes.}

\item{min_time}{Numeric. Minimum daily dive time in minutes. From
\link{dive_pars}.}

\item{round_hr}{Numeric. Round up to nearest portion of an hour. From
\link{dive_pars}.}
}
\value{
Numeric. Adjusted dive time in hours.
}
\description{
Adjust daily dive time according to the collective agreement for biologists
(see \link{dive_pars}).
}
\examples{
adjust_time(dive_time = 12)
adjust_time(dive_time = 122)
}
\seealso{
\link{dive_pars}

Other calculations functions: 
\code{\link{calc_daily_hrs}()},
\code{\link{calc_period}()},
\code{\link{calc_raw_mins}()}
}
\concept{calculations functions}
