% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BubbleTime.R
\docType{package}
\name{BubbleTime-package}
\alias{BubbleTime}
\alias{BubbleTime-package}
\title{BubbleTime: Calculate Dive Time.}
\description{
Calculate adjusted daily dive time based on dive start and end
times according to the collective agreement for biologists. Dive time may be
used to calculate dive pay; double-check dive times before submitting them.
}
\details{
The BubbleTime package provides example data, parameters, as well as four
families of functions: loaders, calculations, figures, and helpers.
Calculations are based on the collective agreement for biologists (see
\link{dive_pars}).
}
\section{Data}{
 The data are dive times (\link{dives}) and dive parameters
(\link{dive_pars}).
}

\section{Loaders functions}{
 The loaders functions are \code{\link[=load_dives]{load_dives()}}.
}

\section{Calculations functions}{
 The calculations functions are
\code{\link[=calc_raw_mins]{calc_raw_mins()}}, \code{\link[=adjust_time]{adjust_time()}}, \code{\link[=calc_period]{calc_period()}}, and
\code{\link[=calc_daily_hrs]{calc_daily_hrs()}}.
}

\section{Figures functions}{
 The figures functions are \code{\link[=plot_daily_mins]{plot_daily_mins()}} and
\code{\link[=plot_cumulative_hrs]{plot_cumulative_hrs()}}.
}

\section{Helpers functions}{
 The helpers functions are \code{\link[=choose_file]{choose_file()}} and
\code{\link[=dive_buddy]{dive_buddy()}}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/grinnellm/BubbleTime}
  \item Report bugs at \url{https://github.com/grinnellm/BubbleTime/issues}
}

}
\author{
\strong{Maintainer}: Matthew H. Grinnell \email{matthew.grinnell@dfo-mpo.gc.ca}

}
\keyword{internal}
