% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_daily_hrs}
\alias{calc_daily_hrs}
\title{Calculate adjusted daily dive time.}
\usage{
calc_daily_hrs(dat, ...)
}
\arguments{
\item{dat}{Tibble. From \code{\link[=calc_raw_mins]{calc_raw_mins()}}.}

\item{...}{Other arguments passed to \code{\link[=adjust_time]{adjust_time()}}.}
}
\value{
Tibble with Date and one column for each diver indicating daily
adjusted dive time in hours. This can be used to calculate dive pay.
}
\description{
Calculate adjusted daily dive time in hours using raw minutes for each dive
and rules in \code{\link[=adjust_time]{adjust_time()}}.
}
\note{
Input is minutes for each dive; output is daily hours for each diver.
}
\examples{
raw_mins <- calc_raw_mins(dat = dives)
calc_daily_hrs(dat = raw_mins)
}
\seealso{
\code{\link[=calc_raw_mins]{calc_raw_mins()}} \code{\link[=adjust_time]{adjust_time()}}

Other calculations functions: 
\code{\link{adjust_time}()},
\code{\link{calc_period}()},
\code{\link{calc_raw_mins}()}
}
\concept{calculations functions}
