% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_raw_mins}
\alias{calc_raw_mins}
\title{Calculate raw dive time.}
\usage{
calc_raw_mins(dat, ...)
}
\arguments{
\item{dat}{Tibble. From \code{\link[=load_dives]{load_dives()}}.}

\item{...}{Other arguments passed to \code{\link[=calc_period]{calc_period()}}.}
}
\value{
Tibble with four columns: Date, Transect, Diver, and Time (in
minutes).
}
\description{
Calculate raw dive time in minutes for each dive based on start and end
times.
}
\note{
Input is time; output is minutes.
}
\examples{
calc_raw_mins(dat = dives)
}
\seealso{
\code{\link[=load_dives]{load_dives()}}

Other calculations functions: 
\code{\link{adjust_time}()},
\code{\link{calc_daily_hrs}()},
\code{\link{calc_period}()}
}
\concept{calculations functions}
