% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dive_pars}
\alias{dive_pars}
\title{Parameters for dive time calculations.}
\format{
List with items:
\describe{
\item{min_time}{Minimum dive time in minutes.}
\item{round_hr}{Round up to nearest portion of an hour (i.e., 0.25 rounds
up to the nearest 1/4 hour).}
\item{dive_period}{Period of time in hours that is equal to one dive day.}
}
}
\usage{
dive_pars
}
\description{
Parameters for daily dive time calculations according to the collective
agreement for biologists (see notes).
}
\note{
Relevant section in the \href{https://www.tbs-sct.canada.ca/agreements-conventions/view-visualiser-eng.aspx?id=3#tocxx245084}{collective agreement for biologists}.

\strong{Article 22: diving allowance}

\strong{22.01} Employees whose job duties require them to dive (as that word is
hereinafter defined) shall be paid an extra allowance of twenty-five
dollars ($25) per hour. The minimum allowance shall be two (2) hours per
dive.

\strong{22.02} A dive is the total of any period or periods of time during any
eight (8) hour period in which an employee carries out required underwater
work with the aid of a self-contained air supply.
}
\examples{
dive_pars
}
\seealso{
\code{\link[=adjust_time]{adjust_time()}} \code{\link[=calc_period]{calc_period()}}

Other data: 
\code{\link{dives}},
\code{\link{dives_simple}}
}
\concept{data}
\keyword{datasets}
