% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dives}
\alias{dives}
\title{Dive times.}
\format{
Tibble with rows for transects and columns for divers.
}
\usage{
dives
}
\description{
Table with example dive times. The first column is date (YYYY-MM-DD), the
second column is transect. Other columns are start and end times (HH:MM) in
24-hour clock for each diver. Column names must be in the format "Matt.Start"
and "Matt.End."
}
\note{
Dives that span midnight must be entered as two separate dives: the
first part of the dive (before midnight) on the first day and the second
part (after midnight) on the next day. See dive times for Transect 119.
}
\examples{
dives
}
\seealso{
\code{\link[=load_dives]{load_dives()}}

Other data: 
\code{\link{dive_pars}}
}
\concept{data}
\keyword{datasets}
