% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dives_simple}
\alias{dives_simple}
\title{Dive times (simple).}
\format{
Tibble with rows for transects.
}
\usage{
dives_simple
}
\description{
Table with example dive times (simplified). The first column is date
(YYYY-MM-DD), the second column is transect, the third column is diver, and
the last two columns are start and end times.
}
\note{
This is for internal use/tests only; do not use this format to
calculate dive time on a survey.
}
\examples{
dives_simple
}
\seealso{
\code{\link[=calc_period]{calc_period()}}

Other data: 
\code{\link{dive_pars}},
\code{\link{dives}}
}
\concept{data}
\keyword{datasets}
