% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{load_dives}
\alias{load_dives}
\title{Load dive times from file.}
\usage{
load_dives(file_path)
}
\arguments{
\item{file_path}{Path and name of a CSV file with dive time information. See
\link{dives} for more information.}
}
\value{
Tibble with rows for transects and columns for diver times.
}
\description{
Load a table of dive times from a CSV file. See the help file for the \link{dives}
data for required format (i.e., column names and types).
}
\note{
Dives that span midnight must be entered as two separate dives: the
first part of the dive (before midnight) on the first day and the second
part (after midnight) on the next day.
}
\examples{
my_dives <- load_dives(
  file_path = system.file("extdata", "dives.csv", package = "BubbleTime")
)
\dontrun{
my_dives <- load_dives(file_path = choose_file())
}
}
\seealso{
\link{dives} \code{\link[=choose_file]{choose_file()}}
}
\concept{loaders functions}
