% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_cumulative_hrs}
\alias{plot_cumulative_hrs}
\title{Plot cumulative dive time.}
\usage{
plot_cumulative_hrs(dat)
}
\arguments{
\item{dat}{Tibble. From \code{\link[=calc_daily_hrs]{calc_daily_hrs()}}.}
}
\value{
Time series of cumulative dive time.
}
\description{
Plot cumulative adjusted daily dive time in hours for each diver.
}
\examples{
raw_mins <- calc_raw_mins(dat = dives)
daily_hrs <- calc_daily_hrs(dat = raw_mins)
plot_cumulative_hrs(dat = daily_hrs)
}
\seealso{
\code{\link[=calc_daily_hrs]{calc_daily_hrs()}} \code{\link[=dive_buddy]{dive_buddy()}}

Other figures functions: 
\code{\link{plot_daily_mins}()}
}
\concept{figures functions}
