% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_daily_mins}
\alias{plot_daily_mins}
\title{Plot daily dive time.}
\usage{
plot_daily_mins(dat)
}
\arguments{
\item{dat}{Tibble. From \code{\link[=calc_raw_mins]{calc_raw_mins()}}.}
}
\value{
Plot with bars for each diver and facets for each day.
}
\description{
Plot raw daily dive time in minutes for each diver based on start and end
times.
}
\examples{
raw_mins <- calc_raw_mins(dat = dives)
plot_daily_mins(dat = raw_mins)
}
\seealso{
\code{\link[=calc_raw_mins]{calc_raw_mins()}} \code{\link[=dive_buddy]{dive_buddy()}}

Other figures functions: 
\code{\link{plot_cumulative_hrs}()}
}
\concept{figures functions}
