% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_period}
\alias{calc_period}
\title{Determine dive period.}
\usage{
calc_period(dat, dive_period = dive_pars$dive_period)
}
\arguments{
\item{dat}{Tibble. Columns for date, diver, as well as start and end times
for each dive. See \code{\link[=calc_raw_mins]{calc_raw_mins()}}.}

\item{dive_period}{Numeric. Number of hours that constitutes a dive period.
From \link{dive_pars}.}
}
\value{
Numeric. Dive period indicating if the dive is in the first (1),
second (2), or third (3) period of the day where periods start at the start
of the first dive of the day.
}
\description{
Determine the dive period according to the collective agreement for
biologists (see \link{dive_pars}).
}
\examples{
calc_period(dat = dives_simple)
}
\seealso{
\link{dive_pars}

Other calculations functions: 
\code{\link{adjust_time}()},
\code{\link{calc_daily_hrs}()},
\code{\link{calc_raw_mins}()}
}
\concept{calculations functions}
